MLIBZ-1373 Remove Client, File and User tests from UnitTestFramework.  Make sure that all tests from these files are in Tests.Integration
